# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Aura Vault — REST API
  version: "1.0.0"
  description: |
    Backend REST API for the Aura Vault Protocol.

    ## Authentication
    All protected endpoints require a Bearer JWT access token obtained from
    `POST /api/auth/login`.  Access tokens expire after **15 minutes**.
    Use `POST /api/auth/refresh` to obtain a new pair without re-authenticating.

    ## Rate Limiting
    Auth endpoints (`/api/auth/*`) are limited to **20 requests per 15 minutes**
    per IP address.  Exceeding this returns `429 Too Many Requests`.

    ## Base URL
    | Environment | Base URL |
    |---|---|
    | Local | `http://localhost:3001` |
    | Testnet | `https://api-testnet.auravault.xyz` |
    | Mainnet | `https://api.auravault.xyz` |
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api-testnet.auravault.xyz
    description: Stellar Testnet
  - url: https://api.auravault.xyz
    description: Stellar Mainnet
...
paths:
...
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: |
        Exchange a valid refresh token for a new access + refresh token pair.
        The old refresh token is invalidated (token rotation).
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
            example:
              refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        "200":
          description: New token pair issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `POST /api/auth/login`.
        Pass as `Authorization: Bearer <token>`.
...
  schemas:
...
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          description: A valid, non-expired refresh token.
...
    TokenPair:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken:
          type: string
          description: Short-lived JWT for authenticating API requests (15 min).
        refreshToken:
          type: string
          description: Long-lived JWT for obtaining new access tokens (30 days).
        expiresIn:
          type: integer
          description: Access token lifetime in seconds.
          example: 900
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
          example: "walletAddress required"
  responses:
    BadRequest:
      description: Invalid or missing request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "walletAddress required"
    Unauthorized:
      description: Missing, invalid, or expired token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            MissingToken:
              value:
                error: "Missing token"
            InvalidToken:
              value:
                error: "Invalid or expired token"
            ExpiredRefresh:
              value:
                error: "Invalid or expired refresh token"
    RateLimited:
      description: Too many requests — auth rate limit exceeded (20 req / 15 min)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Too many auth requests, try again later"
...